+++ lib/core/prelude/i18n/const.d.ts
@@ -24,12 +24,3 @@ export declare const locale: Locale;
  * The default application region
  */
 export declare const region: RegionStore;
-/**
- * A dictionary to map literal pluralization forms to numbers
- */
-export declare const pluralizeMap: Pick<{
-    none: number;
-    one: number;
-    some: number;
-    many: number;
-}, "some" | "none" | "one" | "many">;