+++ lib/lang/interface.d.ts
@@ -9,6 +9,13 @@ export declare type LangPacs = {
     [key in Language]?: KeysetTranslations;
 };
 export declare type Translation = string | PluralTranslation;
-export declare type PluralTranslation = [one: string, some: string, many: string, none: string];
+export interface PluralTranslation {
+    one: string;
+    two?: string;
+    few?: string;
+    many?: string;
+    zero?: string;
+    other?: string;
+}
 export declare type Translations = Dictionary<Translation>;
 export declare type KeysetTranslations = Dictionary<Translations>;